Pin the EQ's drawn curve and its audio to one golden, fix both drifts
BlockEq.cpp and its TypeScript mirror eqMath.ts promised that the drawn
curve is the audio truth, but nothing checked it. Both are now pinned to
test/files/eq_response_golden.json: BlockEqGoldenTest drives the real
filter with sines and measures it, eqMath.test.ts evaluates the mirror.

The golden found two real defects once it covered the oversampled chain
rates the EQ actually runs at (48 kHz x1 to x8):

- BlockEq stored biquad coefficients and state as float. At 384 kHz a
  low band's poles crowd z = 1 and rounding moved a 20 Hz low cut by up
  to 2.5 dB. Coefficients and state are now double; a recursive
  per-sample filter can't vectorize, so this costs nothing.
- eqMath evaluated |H|^2 as an expanded cosine polynomial whose terms
  cancel down to ~w^4, a 0.1 dB drawing error at x8. It now evaluates
  H(e^jw) as a complex ratio, which only cancels to ~w^2.

Adds vitest (npm test) with a tsconfig.test.json so tests are typechecked
by tsc -b but never bundled, plus BlockEq unit tests for band-role
coercion, clamping, bypass and ValueTree round trips.

+++ plugin/src/BlockEq.cpp
@@ -129,7 +129,8 @@ BlockEq::Band BlockEq::clampBand(Band band) {
 }
 
 // RBJ Audio EQ Cookbook coefficients, A = 10^(dB/40). Keep in exact sync with
-// the TypeScript mirror in ui/src/components/eqMath.ts.
+// the TypeScript mirror in ui/src/components/eqMath.ts; both are pinned to
+// test/files/eq_response_golden.json (BlockEqGoldenTest / eqMath.test.ts).
 void BlockEq::updateBand(int index) {
   const Band& band = bands[static_cast<size_t>(index)];
   Biquad& f = filters[static_cast<size_t>(index)];
@@ -191,11 +192,11 @@ void BlockEq::updateBand(int index) {
   }
 
   const double norm = 1.0 / a0;
-  f.b0 = static_cast<float>(b0 * norm);
-  f.b1 = static_cast<float>(b1 * norm);
-  f.b2 = static_cast<float>(b2 * norm);
-  f.a1 = static_cast<float>(a1 * norm);
-  f.a2 = static_cast<float>(a2 * norm);
+  f.b0 = b0 * norm;
+  f.b1 = b1 * norm;
+  f.b2 = b2 * norm;
+  f.a1 = a1 * norm;
+  f.a2 = a2 * norm;
 }
 
 void BlockEq::updateActivity() {
